Install azure-cli with pip
Installing it with apt-get was causing issues when running some commands
that need uamqp library.

Note that since the python docker image already has `setuptools== 58.1.0`,
we needed to downgrade to a version < 58.0.0 to make it compatible with
azure-cli==2.25.0 (compatible with azure-cli-core==2.25.0 pinned
in this project). See related [issue](https://github.com/Azure/azure-cli/issues/19468).