Update azure-cli-core
The azure-cli-core dependency is currently pinned to version 2.25.0,
which is quite old, so it should be updated. Apart from that, if we want
to install azure-cli when within the same environment using pip, the
same version of azure-cli needs to be used. This version has a
dependency on an old library that makes use of a deprecated feature of
setuptools. Therefore, if we want to set the development environment
this way, we need to use an old setuptools library.

azure-cli-core should be pinned to version 2.34.1. Please note >=2.35.0
cannot be used as this version is not compatible with the docker
dependency, since docker requires websocket-client==0.56.0 and
azure-cli-core>=2.35.0 requires websocket-client==1.31.1.

Please note we removed the manifest for no_modules and the related test
cases because the validation of iot edge deployment create got stricter
with the current version and otherwise the test failed.

Fixes: #560
Co-authored-by: Isabel Lopez Andrade <beandrad@microsoft.com>