Deprecate duplicate commands
Some commands can be called in two different ways (for instance,
`iotedgedev init` and `iotedgev solution init`). Eventually, this
duplication should be removed to simplify the tool. This is a breaking
change, though; so before implementing it in the next major release, we
should add a deprecation notice in such commands. This deprecation
notice should be added to the help message and printed when the command
is run. This can be done by setting `deprecated=True` in the
`click.Command()` decorator (see
[documentation](https://click.palletsprojects.com/en/8.1.x/api/#click.Command)).

Note that the reason for this duplication is that the commands are added
to two `click` groups (`main` and `solution`). The commands that should be
deprecated are the ones corresponding to the `main` group.

Fix: #559